For_Com/requirements.txt
pycryptodome==3.18.0
pyserial==3.5
